Refactor `WP_Config` namespace and introduce `UM()->admin()->wp_config()` helper for improved maintainability and consistency
