+++ composer.json
@@ -43,13 +43,7 @@
       "@php -r \"$dest = 'includes/lib/php-scoper-wordpress-excludes/'; if (!is_dir($dest)) { mkdir($dest, 0755, true); }; copy('vendor/sniccowp/php-scoper-wordpress-excludes/generated/exclude-wordpress-functions.json', $dest . 'exclude-wordpress-functions.json');\""
     ],
     "wp-config-transformer": [
-      "@php -r \"@mkdir('includes/lib/wp-config-transformer', 0755, true); if (is_file('vendor/wp-cli/wp-config-transformer/src/WPConfigTransformer.php')) { copy('vendor/wp-cli/wp-config-transformer/src/WPConfigTransformer.php', 'includes/lib/wp-config-transformer/WPConfigTransformer.php'); }\""
-    ],
-    "post-install-cmd": [
-      "@wp-config-transformer"
-    ],
-    "post-update-cmd": [
-      "@wp-config-transformer"
+      "@php -r \"@mkdir('includes/lib/wp-config-transformer/src', 0755, true); if (is_file('vendor/wp-cli/wp-config-transformer/src/WPConfigTransformer.php')) { copy('vendor/wp-cli/wp-config-transformer/src/WPConfigTransformer.php', 'includes/lib/wp-config-transformer/src/WPConfigTransformer.php'); }\""
     ]
   },
   "extra": {
